Move everything to the energy subpackage
